reconfiguring for vercel
